requirements.txt
certifi==2018.4.16
future==0.16.0
Pillow==5.2.0
python-telegram-bot==10.1.0
pyzbar==0.1.7
redis==2.10.6
